Fixed the AppNew.vue layout
in order to fix the layout for this component, i had to give the template tag a class of row. I did as well change the default size of my app to be fairly compact with a 1000px width. With this it is still flexible in relation of the window size.